requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# Async & WebSockets
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Validation & Serialization
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.27.0

# Rate Limiting
django-ratelimit==4.1.0

# Caching
django-redis==5.4.0

# Task Queue (for background jobs)
celery==5.3.4
redis==5.0.1

# Utilities
uuid==1.30
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
factory-boy==3.3.0
faker==22.0.0

# Code Quality
flake8==7.0.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7

# Production
gunicorn==21.2.0
uvicorn==0.25.0
whitenoise==6.6.0


